requirements.txt
# Framework Web y Extensiones
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.2
Flask-Limiter==2.3.0  # Nuevo: para rate limiting

# Base de Datos
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3  # Para PostgreSQL
PyMySQL==1.0.2  # Para MySQL

# Autenticación y Seguridad
PyJWT==2.3.0
cryptography==36.0.1
python-dotenv==0.19.2

# DocuSign
docusign-esign==3.17.0
requests==2.27.1

# Testing y Desarrollo
pytest==7.1.2
pytest-cov==3.0.0
black==22.3.0
flake8==4.0.1
pre-commit==2.19.0

# Utilidades
python-dateutil==2.8.2
pytz==2022.1
urllib3==1.26.8
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
prometheus_client==0.13.1  # Nuevo: para monitoreo en tiempo real
bleach==4.1.0        # Nuevo: para sanitización de datos

# PDF Generation
reportlab==4.0.4
